pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ledger"
version = "0.2.0"
description = "Schema registry and data obligation manager"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Jeremy McEntire", email = "jandrewmcentire@gmail.com" },
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
]
mock = [
    "faker>=22.0",
]
postgres = [
    "psycopg2-binary>=2.9",
]
redis = [
    "redis>=5.0",
]
mongo = [
    "pymongo>=4.0",
]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[project.scripts]
ledger = "cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/config", "src/registry", "src/migration", "src/export", "src/mock", "src/cli", "src/api", "src/inference"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "--import-mode=importlib"
testpaths = ["tests"]
